1. Displayed current weather data from api 2. Implemented api to get weather forecasts

+++ app/src/main/java/com/spacelabs/weatherapp/domain/Coord.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 public class Coord {
 
-    private int lon;
-    private int lat;
+    private float lon;
+    private float lat;
 
 }
